+++ src/components/Controllers/Text/Editor.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
-import Columns from '../../Parts/Columns'
+import Columns from '../../Parts/Grids/Columns'
 import SpinnerCard from '../../Parts/Cards/Spinner'
 import TextEditor from '../../Editors/Text'
 import { useStore, selectors } from './store'
