+++ src/components/Editors/Base/Item.tsx
@@ -4,7 +4,7 @@ import Box from '@mui/material/Box'
 import Link from '@mui/material/Link'
 import Button from '@mui/material/Button'
 import Typography from '@mui/material/Typography'
-import Columns from '../../Parts/Columns'
+import Columns from '../../Parts/Grids/Columns'
 import HeadingBox from './Heading/Box'
 
 export interface EditorItemProps {
