+++ src/components/Editors/Base/List.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import startCase from 'lodash/startCase'
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
-import Columns from '../../Parts/Columns'
+import Columns from '../../Parts/Grids/Columns'
 import EditorListItem from './ListItem'
 import HeadingBox from './Heading/Box'
