+++ src/components/Editors/Base/Section.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
-import Columns from '../../Parts/Columns'
+import Columns from '../../Parts/Grids/Columns'
 import HeadingBox from './Heading/Box'
 
 export interface EditorItemProps {
