+++ src/components/Editors/Chart/Layout.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import camelCase from 'lodash/camelCase'
 import { useTheme } from '@mui/material/styles'
 import Box from '@mui/material/Box'
-import Columns from '../../Parts/Columns'
+import Columns from '../../Parts/Grids/Columns'
 import VerticalTabs from '../../Parts/Tabs/Vertical'
 import EditorHelp from '../Base/Help'
 import Chart from './Sections/Chart/Chart'
