+++ src/components/Editors/Chart/Sections/Chart/Chart.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box'
 import InputField from '../../../../Parts/Fields/Input'
 import SelectField from '../../../../Parts/Fields/Select'
 import EditorSection from '../../../Base/Section'
-import Columns from '../../../../Parts/Columns'
+import Columns from '../../../../Parts/Grids/Columns'
 import { useStore, selectors } from '../../store'
 import * as settings from '../../settings'
 import YesNoField from '../../../../Parts/Fields/YesNo'
