+++ src/components/Editors/Chart/Sections/Chart/LayerChart.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import Box from '@mui/material/Box'
 import SelectField from '../../../../Parts/Fields/Select'
 import EditorSection from '../../../Base/Section'
-import Columns from '../../../../Parts/Columns'
+import Columns from '../../../../Parts/Grids/Columns'
 import { useStore } from '../../store'
 import * as settings from '../../settings'
