+++ src/components/Editors/Dialect/Sections/Dialect.tsx
@@ -4,7 +4,7 @@ import InputField from '../../../Parts/Fields/Input'
 import YesNoField from '../../../Parts/Fields/YesNo'
 import MultilineField from '../../../Parts/Fields/Multiline'
 import EditorSection from '../../Base/Section'
-import Columns from '../../../Parts/Columns'
+import Columns from '../../../Parts/Grids/Columns'
 import * as settings from '../../../../settings'
 import { useStore } from '../store'
 import validator from 'validator'
