+++ src/components/Editors/Dialect/Sections/Html.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
-import Columns from '../../../Parts/Columns'
+import Columns from '../../../Parts/Grids/Columns'
 import InputField from '../../../Parts/Fields/Input'
 import EditorSection from '../../Base/Section'
 import * as settings from '../../../../settings'
