+++ src/components/Editors/Package/Layout.tsx
@@ -7,7 +7,7 @@ import HorizontalTabs from '../../Parts/Tabs/Horizontal'
 import VerticalTabs from '../../Parts/Tabs/Vertical'
 import EditorHelp from '../Base/Help'
 import SelectField from '../../Parts/Fields/Select'
-import Columns from '../../Parts/Columns'
+import Columns from '../../Parts/Grids/Columns'
 import Resource from '../Resource'
 import Dialect from '../Dialect'
 import Schema from '../Schema'
