+++ src/components/Editors/Schema/Sections/Field.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import partition from 'lodash/partition'
 import Box from '@mui/material/Box'
 import { useTheme } from '@mui/material/styles'
-import Columns from '../../../Parts/Columns'
+import Columns from '../../../Parts/Grids/Columns'
 import EditorItem from '../../Base/Item'
 import EditorList from '../../Base/List'
 import EditorListItem from '../../Base/ListItem'
