+++ src/components/Parts/Bars/Action.tsx
@@ -7,7 +7,7 @@ import SaveAltIcon from '@mui/icons-material/SaveAlt'
 import CheckIcon from '@mui/icons-material/Check'
 import HistoryIcon from '@mui/icons-material/History'
 import IconButton from '../../Parts/Buttons/Icon'
-import Columns from '../../Parts/Columns'
+import Columns from '../Grids/Columns'
 import LightTooltip from '../Tooltips/Light'
 import { useKeyPress } from 'ahooks'
